requirements.txt
Flask==3.0.3
flask-cors==4.0.1
pyTelegramBotAPI==4.21.0
python-dotenv==1.0.1
pyotp==2.9.0
qrcode==7.4.2
pypng